feat: extend nodeByUri, support custom fields
